SFT-2176: moved fcc_copy_files_task to developer folder
