+++ ports/stm32/boards/Passport/modules/tasks/__init__.py
@@ -26,7 +26,6 @@ from .delete_multisig_task import delete_multisig_task
 from .double_check_psbt_change_task import double_check_psbt_change_task
 from .erase_passport_task import erase_passport_task
 from .format_microsd_task import format_microsd_task
-# from .fcc_copy_files_task import fcc_copy_files_task
 from .generate_addresses_task import generate_addresses_task
 from .get_security_words_task import get_security_words_task
 from .get_seed_words_task import get_seed_words_task
